Separate metadata interceptor into new module
