+++ tests/interceptors/metadata_interceptor_test.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Tests for the Metadata gRPC Interceptor"""
+"""Tests for the Metadata gRPC Interceptor."""
 
 import mock
 from unittest import TestCase
